Add read-through annotation caching and re-generation
This fixes a bug when Gedmo-specific annotation caches are dropped without dropping the related Doctrine annotation. This forces Gedmo metadata re-generation on a cache miss. In order to prevent useless annotation parsing, an empty cache entries are created for Doctrine extensions which are enabled but not present on a given class.

Fixes #89